backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Celery
celery[redis]==5.4.0

# HTTP Client
httpx==0.28.1

# Image Processing
Pillow==11.1.0

# AWS S3 / Cloudflare R2
boto3==1.36.2

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# Utilities
python-slugify==8.0.4
python-dateutil==2.9.0

# Security
cryptography==44.0.0

# CORS
starlette==0.41.3

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Linting
ruff==0.9.2
